+++ src/main/java/com/eeehit/board/controller/AdminController.java
@@ -1,0 +1,34 @@
+package com.eeehit.board.controller;
+
+import com.eeehit.board.domain.Role;
+import com.eeehit.board.domain.User;
+import com.eeehit.board.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+/**
+ * Created by trinity on 17. 5. 18.
+ */
+@RestController
+@RequestMapping("admin" )
+public class AdminController {
+    @Autowired
+    private UserService userService;
+
+    @RequestMapping(value = "userList", method = RequestMethod.GET)
+    public List<User> getUserList(){
+        return this.userService.getUserList();
+    }
+
+    @RequestMapping(value = "userId/{userId}", method = RequestMethod.GET)
+    public User getUserByUserId(@PathVariable(value = "userId", required = true) String userId){
+        return this.userService.getUserByUserId(userId);
+    }
+
+    @RequestMapping(value = "join", method = RequestMethod.POST)
+    public User join(@RequestParam(value = "userId", required = true) String userId, @RequestParam(value = "userPw", required = true) String userPw) {
+        return this.userService.join(userId, userPw, Role.USER);
+    }
+}
