+++ src/main/java/com/eeehit/board/controller/UserController.java
@@ -1,34 +1,31 @@
 package com.eeehit.board.controller;
 
-import com.eeehit.board.domain.Role;
 import com.eeehit.board.domain.User;
 import com.eeehit.board.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.servlet.ModelAndView;
 
-import java.util.List;
+import javax.servlet.http.HttpSession;
 
 /**
- * Created by trinity on 17. 5. 18.
+ * Created by trinity on 17. 6. 12.
  */
-@RestController
-@RequestMapping("admin" )
+@Controller
+@RequestMapping("myPage")
 public class UserController {
     @Autowired
-    private UserService userService;
+    UserService userService;
 
-    @RequestMapping(value = "userList", method = RequestMethod.GET)
-    public List<User> getUserList(){
-        return this.userService.getUserList();
-    }
 
-    @RequestMapping(value = "userId/{userId}", method = RequestMethod.GET)
-    public User getUserByUserId(@PathVariable(value = "userId", required = true) String userId){
-        return this.userService.getUserByUserId(userId);
-    }
-
-    @RequestMapping(value = "join", method = RequestMethod.POST)
-    public User join(@RequestParam(value = "userId", required = true) String userId, @RequestParam(value = "userPw", required = true) String userPw) {
-        return this.userService.join(userId, userPw, Role.USER);
+    @RequestMapping(value = "", method = RequestMethod.GET)
+    public ModelAndView myPage(HttpSession httpSession){
+        ModelAndView mv = new ModelAndView("myPage");
+        String userId = (String) httpSession.getAttribute("userId");
+        User user = userService.getUserByUserId(userId);
+        mv.addObject("user", user);
+        return mv;
     }
 }
