+++ src/main/resources/templates/user/myPage.hbs
@@ -1,0 +1,16 @@
+<div class="container">
+    <div class="row">
+        <div class="col-lg-8 col-lg-offset-2 col-md-10 col-md-offset-1">
+            {{#user}}
+                <div class="post">
+                    <h2>
+                        {{userId}}
+                    </h2>
+                    <h3>
+                        {{role}}
+                    </h3>
+                </div>
+            {{/user}}
+        </div>
+    </div>
+</div>
